requirements.txt
alembic==1.4.0
bcrypt==3.1.7
bleach==3.1.1
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
Click==7.0
cloudinary==1.24.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Markdown==0.3
Flask-Migrate==2.5.2
Flask-MySQL==1.5.1
Flask-Sijax==0.4.1
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
getmac==0.8.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
Markdown==3.2.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
netifaces==0.10.9
Pillow==7.0.0
protobuf==3.15.3
psycopg2-binary==2.8.6
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
regex==2020.2.20
requests==2.25.1
Sijax==0.3.2
six==1.14.0
slugify==0.0.1
SQLAlchemy==1.3.13
text-unidecode==1.3
timeago==1.0.15
Unidecode==1.1.1
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.0
wtf-tinymce==0.1.3
WTForms==2.2.1
